+++ packages/sealed_countries/lib/src/helpers/world_country/world_country_names.dart
@@ -58,6 +58,7 @@ extension WorldCountryNames on WorldCountry {
 
     return _nameTranslated(language);
   }
+  // TODO! Remove.
 
   /// Returns the Arabic name of the country.
   @Deprecated(