+++ packages/world_countries/lib/src/interfaces/searchable_interface.dart
@@ -1,3 +1,5 @@
+import "package:flutter/widgets.dart" show BuildContext;
+
 /// An interface for classes that can be searched.
 abstract interface class SearchableInterface<T extends Object?> {
   /// Constructor for the [SearchableInterface] class.
@@ -17,7 +19,7 @@ abstract interface class SearchableInterface<T extends Object?> {
 
   /// A function that takes an item and returns an iterable of strings to search
   /// in.
-  final Iterable<String> Function(T item)? searchIn;
+  final Iterable<String> Function(T item, BuildContext context)? searchIn;
 
   /// A boolean indicating whether the search should be case-sensitive.
   final bool caseSensitiveSearch;