+++ packages/world_countries/lib/src/models/typedefs.dart
@@ -1,0 +1,17 @@
+import "dart:ui" show Locale;
+
+import "locale/typed_locale.dart";
+
+/// A typedef representing a map entry of `Locale` and `TypedLocale`.
+///
+///
+/// The `LocaleEntry` typedef is a shorthand for `MapEntry<Locale, TypedLocale>`
+/// type. It defines a map entry where the key is of type [Locale] and
+/// the value is of type [TypedLocale].
+///
+/// Example usage:
+///
+/// ```dart
+/// const localeEntry = LocaleEntry(Locale('en'), TypedLocale(LangEng()));
+/// ```
+typedef LocaleEntry = MapEntry<Locale, TypedLocale>;