+++ packages/world_countries/lib/src/widgets/base_widgets/stateful_searchable.dart
@@ -25,7 +25,7 @@ abstract class StatefulSearchable<T extends Object> extends StatefulWidget
   final TextEditingController textController;
 
   @override
-  final Iterable<String> Function(T item) searchIn;
+  final Iterable<String> Function(T item, BuildContext context) searchIn;
 
   @override
   final bool caseSensitiveSearch;