+++ packages/world_countries/lib/src/widgets/country/country_tile.dart
@@ -1,6 +1,7 @@
 import "package:flutter/material.dart";
 import "package:sealed_countries/sealed_countries.dart";
 
+import "../../constants/ui_constants.dart";
 import "../../models/item_properties.dart";
 import "../generic_widgets/list_item_tile.dart";
 import "emoji_flag.dart";
@@ -58,8 +59,6 @@ class CountryTile extends ListItemTile<WorldCountry> {
   ///   display information about.
   /// * All other parameters are optional and are passed to the superclass
   ///   constructor.
-  /// * [translation] is an optional `NaturalLanguage` object used to translate
-  ///   the country's name.
   CountryTile.fromProperties(
     ItemProperties<WorldCountry> country, {
     Widget? leading,
@@ -95,21 +94,20 @@ class CountryTile extends ListItemTile<WorldCountry> {
     super.titleAlignment,
     super.titleTextStyle,
     super.visualDensity,
-    NaturalLanguage? translation,
   }) : super(
           country.item,
           isChosen: country.isChosen,
           isDisabled: country.isDisabled,
-          title: title ?? Text(country.item.namesNative.first.common),
-          leading: leading ?? EmojiFlag.twemoji(country.item),
-          subtitle: subtitle ??
+          title: title ??
               Text(
-                country.item
-                        .maybeTranslation(translation ?? const LangEng())
-                        ?.name ??
-                    country.item.namesCommonNative(skipFirst: true),
+                country.item.namesCommonNative(skipFirst: true),
                 overflow: TextOverflow.ellipsis,
               ),
+          leading: Padding(
+            padding: const EdgeInsets.only(top: UiConstants.point),
+            child: leading ?? EmojiFlag.twemoji(country.item),
+          ),
+          subtitle: subtitle ?? Text(country.item.namesNative.first.common),
         );
 
   /// Constructor for the [CountryTile] class that uses an [ItemProperties]
@@ -119,8 +117,6 @@ class CountryTile extends ListItemTile<WorldCountry> {
   ///   display information about.
   /// * All other parameters are optional and are passed to the superclass
   ///   constructor.
-  /// * [translation] is an optional `NaturalLanguage` object used to translate
-  ///   the country's name.
   CountryTile.simple(
     ItemProperties<WorldCountry> country, {
     Widget? leading,
@@ -156,17 +152,13 @@ class CountryTile extends ListItemTile<WorldCountry> {
     super.titleAlignment,
     super.titleTextStyle,
     super.visualDensity = VisualDensity.compact,
-    NaturalLanguage? translation,
   }) : super(
           country.item,
           isChosen: country.isChosen,
           isDisabled: country.isDisabled,
           title: title ??
               Text(
-                country.item
-                        .maybeTranslation(translation ?? const LangEng())
-                        ?.name ??
-                    country.item.namesNative.first.common,
+                country.item.namesNative.first.common,
                 overflow: TextOverflow.ellipsis,
               ),
           leading: leading ?? EmojiFlag.twemoji(country.item),